Refactor: 클래스 패키지 리팩토링
- 기존에 구현된 배치작업 클래스를 패키지 성격에 맞게 이동
