+++ src/main/java/com/ssginc/showpingrefactoring/batch/config/VodRecommendBatchConfig.java
@@ -1,8 +1,8 @@
 package com.ssginc.showpingrefactoring.batch.config;
 
 import com.ssginc.showpingrefactoring.domain.stream.dto.object.VodRecommendDto;
-import com.ssginc.showpingrefactoring.batch.job.VodRecommendProcessor;
-import com.ssginc.showpingrefactoring.batch.job.VodRecommendWriter;
+import com.ssginc.showpingrefactoring.batch.job.processor.VodRecommendProcessor;
+import com.ssginc.showpingrefactoring.batch.job.writer.VodRecommendWriter;
 import com.ssginc.showpingrefactoring.batch.listener.JobFailureListener;
 import lombok.RequiredArgsConstructor;
 import org.springframework.batch.core.Job;
